created separate folder for tests inside src
