+++ pokemon-project/src/tests/App.test.js
@@ -1,7 +1,7 @@
 /* eslint-disable testing-library/no-node-access */
 import { render, screen } from "@testing-library/react";
 // import userEvent from "@testing-library/user-event";
-import App from "./App";
+import App from "../App";
 
 test("renders title", () => {
   render(<App />);
