+++ pokemon-project/src/tests/Header.test.js
@@ -1,5 +1,4 @@
-import Header from "./components/Header";
-
+import Header from "../components/Header";
 /* eslint-disable testing-library/no-node-access */
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
